winston and qrcode packages were demonstrated.
winston was used to log into console and file at the same time.
Configuration of the logger was done based on the online documentation.

qucode package was used to generate QR codes with defined text.
We generated URLData which can be loaded into HTML images (src=).
FileSystem was used to save QR Codes to disk.
It requires the fs library of node.